[dune] Use a special directory for the boot build
This is related to and fixes #10694 in part.

When calling bootstrap in an incremental build step, we must be sure
the generated dune files are in place. In the CI, these won't be in
place, so we must bootstrap without altering the digest and trace
database coming from the artifact.

Using a separate boot step to recreate the missing `dune` files works
fine and takes just a few seconds.

(cherry picked from commit bfbcd1e3feeaa7eb24c15686439926c536c5d483)