Use code owner teams for every component.
It was decided during the Coq WG that code owner teams are more
convenient, in particular because they allow adding and removing team
members without going through a pull request.  For each team, we
should aim to have at least three code owners, even if in some cases
we are going to start with less.

We also stop triggering review requests for changelog entries as was
also decided during the WG.