Fixes a small bug exposing an _ANONYMOUS_REL in a unification error message.
Might be improvable further. In the first example, we have two
environments involved and one is implicit. It does not seem excluded
that a variable name of the second environment shows up which is not
listed in the first environment.

(cherry picked from commit 18626b09a8a8ea1d56a7f91f347e1ebb3960e2c6)