[mltop] Store digest of modules used to compile files.
This should allow digest-based builds to work correctly.

Fixes #10874

For now, we store the digest of the module selected, this means that
vo files loading modules will have different digests depending on
whether the `native` or `byte` version was used.

If that is a problem we can improve this tho; for example we could
disable the digest in those cases.

The code could be better, and indeed `Mltop` could enjoy a nice
cleanup, I will likely do some when I add support for `Fl_dynload`.