[ci] [gitlab] Add `interruptible: true` to jobs.
When a newer pipeline contains the same job, the job can be
interrupted, see:

https://docs.gitlab.com/ee/ci/yaml/README.html#interruptible

This should help with job limits (c.f. #11320 )

The patch is a bit unsatisfactory due to the duplication needed;
we could define a base job and use extends, but not sure it is worth
it for now.