Fix 11483
Performance bug of PrimFLoat.compare with native_compute

When adapting Coq.Interval with @erikmd and @silene, we noticed that
PrimFLoat.compare is taking a lot of time with native_compute (much
more than with vm_compute).

This comes from the implementation using the OCaml polymorphic
comparison instead of the float comparison.