[init] Add `-boot` option to avoid binding `Coq.` prefix.
This is useful when we want to have finer control of the location of
files in the bootstrap process, for example when building using Dune.

Also, this makes options consistent with what `coqdep` already uses
for bootstrap.

The main use case for `-boot` is to replace the hardcoded `add_load_path (coqlib () / theories)` with `-R dir Coq`, where dir is controlled by the build system. In particular, we use `-R . Coq` as we `cd` into the directory the package is, so without boot we'd have to hardcode the `theories` path in Dune itself. which seems less robust.

Notably after this change the only part of the build that uses `coqlib` is the micromega solver, but that can be tweaked so if coqlib is not set it will use the one in the path. IMO not having to set "coqlib" is a good property if we want a more compositional setup.