New flag to restrict parsing to the core language.
The use case is pedagogical: we would like the first chapter of the
refactored refman to be about the core language that the kernel
understands, and we would like to be able to give examples where
elaboration plays no role.  This flag will be helpful to ensure that
this is the case.  It will typically be combined with Set Printing
All.