Added the cut function and three auxiliary lemmas, In_inv and NoDup_tail.
The cut function accepts a list and an element in the list and cuts the list at the given element.

The cut function is accompanied by a companion lemma named cut_destr which asserts that the result of the cut function is correct.

In_inv is a computable analogue of in_inv.

NoDup_tail is proves that the tail of a list that does not contain duplicates does not contain any duplicates.