[dune] [ocamldebug] Improve ocamldebug rules
We provide the closure of the dependencies manually.

This is still a hack, but not so bad given that the `source`'d files
still do contain that duplication too.

Dune should provide this functionality so we can replace both this
rule and the source files. Actually, that's not hard to implement,
`utop` already supports a printer attribute so these are loaded
automatically, so the ocamldebug mode could do the same.

Should fix #11716