[API] make declare_variable more uniform wrt declare_axiom
This commit follows 2cdccb3 by making the deprecation path
declare_assumption -> [ declare_variable | declare_axiom ] less error
prone. Before this commit declare_assumption had to be replaced either by
  declare_axioms
of
  Declare.declare_universe_context + declare_variable
which is error prone.

The current API is still unsatisfactory because one has to know if
the universes to be declared are poly or not. At least plugins just
using monomorphic universes can get things right (there is a sanity
check in kerne/section that complain if one passes the wrong poly,
not sure it is worth it).