[obligations] More progress towards unification of the save path
We make internal types `private` as an step towards the unification of
the save path with the rest of the system.

In particular, this is allow us to guarantee invariants
w.r.t. external users as the large majority of fields are always
constant.

This will also enable at some point a common creation of proof entry
with the rest of the system.