Do not rely on the implicit declaration of caml_minor_collection.
This commit also prefixes young_ptr and young_limit along the way, so as
to not rely on OCaml's compatibility layer. This is a gratuitous change,
since this code is only meant to be compiled with OCaml < 4.10 anyway.

(cherry picked from commit 4666a8b9596f8cb87b63c345f6e57348f0bfda6d)