[lemmas] Remove workaround for non-uniform mutual body
In previous versions of Coq `abstract` could create non-uniform bodies
for mutually recursive definitions (c.f. bug #5065)

Thanks to changes to the trusted base this is not the case anymore,
thus we can remove the workaround.

This way mutual bodies are handled the same in the interactive and
non-interactive case.