Nicer printing for decimal constants in R
Print 1.5 as 1.5 and not 15e-1.

We choose the shortest representation
with tie break to the dot notation (0.01 rather than 1e-3).

The printing remains injective, i.e. 12*10^2 is printed 12e2 in order
not to mix it with 1200 and 12/10^1 is not mixed with 120/10^2, the
first being printed as 1.2 and the last as 1.20.