Add a specific opam file to build te docker image
The Docker image coqorg/coq:dev is currently built using the OPAM file
coq.opam. Since this file is used for develoment purpose, it would be
better to use a more stable one for building the docker images. The
OPAM option "--locked=docker" will then be used in the opam pin
command when building the docker image to use the new coq.opam.docker
file.

The new file builds Coq using make, this is temporary and could be
reverted to dune once it supports "-native-compiler yes".