[warnings] [doc] Preliminary support for dumping known warnings
This is very preliminary work adding a `-dump-warnigns` option that
could be used by IDEs and documentation to check exhaustiveness.

There still remains a non-trivial amount of work to do in order to
reify the pretty printing structure so we could produce an structured
representation of the warning message, but it'd be doable.