Temporarily disable Windows job on Azure.
This job is currently broken because of Dune 2.5 not being available
yet on the mingw-opam-repository.

Before this, it was already red most of the time because of two
issues:

- Cygwin cannot be downloaded

- "Error on dynamically loaded library:
.\kernel/byterun\dllbyterun_stubs.dll: The specified module could not
be found" bug."

We don't know exactly when the latter appeared.  It was missed because
of the former.

The goal is to reenable Windows testing based on standard opam with no
cygwin.  Hopefully, this is achievable in the next few weeks.