Notations: Giving a consistent role to global references occurring patterns.
Currently, global references in patterns used also as terms were
accepted for parsing but not for printing.

We accept section variables for both parsing and printing. We reject
constant and inductive types for both parsing and printing.

Among other, this also fixes a hole in interpreting variables used
both patterns and terms: the term part was not interpreted.