Locating error again in TacAtom and TacAlias (fixes #12152, fixes #12255).
Since we don't always have the call trace anymore, we explicitly
insert a catch of failures in TacAlias. The trace is then treated in
this catch rather than propagated to the underlying calls (a VFun?). I
hope this is doing the same.

The suggestion to use a tclOR is from P.-M. Pédrot.

Note: this is not fully ideal, the messages which were expecting a
trace should be rethought to take into account either that the calls
are not printed anymore, or to print them again.

(cherry picked from commit 01e465952e79f4e7fdef3417c6f1883ca2c5e328)