Fix #12211
I forgot to test #12211 sufficiently; it was emitting timing info
without saying which file was being timed, because the evaluation of
`$@` was performed at the definition of `OCAMLC`.  This fixes that issue
of 8bd559370f51d7cc1877380a5ad726da67ceb0fa by delaying the evaluation
of the definitions of `OCAMLC` and `OCAMLOPT` to the running of the
recipies.